add `DataContainer.jl` and delete `TimeLine.jl` #109
